requirements.txt
uvicorn
fastapi
requests
psycopg2-binary